refactor: moved validations to utils
