--- utils/passwordValidation.js
@@ -1,14 +1,0 @@
-const bcrypt = require('bcrypt')
-
-async function hashPassword(password) {
-  return await bcrypt.hash(password, 10)
-}
-
-async function comparePasswords(plainPassword, hashedPassword) {
-  return await bcrypt.compare(plainPassword, hashedPassword)
-}
-
-module.exports = {
-  hashPassword,
-  comparePasswords,
-}
